requirements.txt
Flask==2.3.2
crewai==0.83.0
langchain>=0.2.16
python-dotenv==1.0.0
gunicorn==21.2.0
evolutionapi==0.0.5
mem0ai==0.1.34
psycopg2==2.9.9
langchain-google-community[gmail]==2.0.3
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1